refactor images set creation directory
