+++ backend/app.py
@@ -1,15 +1,14 @@
 from flask import Flask
 from flask_cors import CORS
-from face_extractor.face_extractor_controller import face_extractor_bp
-from game_flow.lobby_controller import lobby_bp
-from game_logic.game_logic_controller import game_logic_bp
+
+from backend.images_set_creation.images_controller import images_bp
+from backend.game_flow.lobby_controller import lobby_bp
 
 app = Flask(__name__)
 CORS(app)  # Enable CORS for cross-origin requests
 
 app.register_blueprint(lobby_bp)
-app.register_blueprint(face_extractor_bp)
-app.register_blueprint(game_logic_bp)
+app.register_blueprint(images_bp, url_prefix='/images')
 
 
 if __name__ == '__main__':
