+++ backend/images_set_creation/face_extractor.py
